% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btls_stn.R
\name{btls_stn}
\alias{btls_stn}
\title{Backtracking line search}
\usage{
btls_stn(x, X, Z, q, dx, u, alpha = 1, rho = 0.5, c = 1e-04, maxit = 20)
}
\arguments{
\item{x}{parameter vector}

\item{X}{design matrix}

\item{Z}{response dummies}

\item{dx}{search direction}

\item{u}{score vector}

\item{alpha}{step size}

\item{rho}{contraction factor}

\item{c}{Armijo parameter}

\item{maxit}{maximum number of iterations}
}
\value{
new parameter vector
}
\description{
Backtracking line search
}
