% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacob.R
\name{jacob}
\alias{jacob}
\title{Jacobian}
\usage{
jacob(alpha, gamma, Lambda, Phi)
}
\arguments{
\item{alpha}{vector of intercepts}

\item{gamma}{vector of regressor effects}

\item{Lambda}{matrix of regressor effects}

\item{Phi}{matrix of weights}
}
\value{
Jacobian matrix
}
\description{
Jacobian
}
