% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{The log-likelihood function}
\usage{
logLik(x, X, Z)
}
\arguments{
\item{x}{parameter vector}

\item{X}{desgin matrix}

\item{Z}{response dummies}
}
\value{
log-likelihood value
}
\description{
The log-likelihood function
}
