% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{lrt}
\alias{lrt}
\title{Likelihood ratio test}
\usage{
lrt(obj1, obj2)
}
\arguments{
\item{obj1}{fit results model 1}

\item{obj2}{fit results model 2}
}
\value{
test statistic, df, and p-value
}
\description{
Likelihood ratio test
}
