% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.R
\name{mlr}
\alias{mlr}
\title{Multinomial logistic regression}
\usage{
mlr(formula, data, maxit = 100)
}
\arguments{
\item{formula}{model expression}

\item{data}{data frame}

\item{maxit}{maximum number of iterations}
}
\value{
A list of
\item{coefficients}{Parameter estimates}
\item{fitted.values}{The model implied probabilities}
\item{residuals}{Dummy responses minus fitted values}
\item{pred.cat}{The model implied categories}
}
\description{
This function can be used to fit the multinomial logistic regression model.
}
\author{
David J. Hessen
}
