% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olr.R
\name{olr}
\alias{olr}
\title{Ordinal logistic regression}
\usage{
olr(formula, data, maxit = 100)
}
\arguments{
\item{formula}{model expression}

\item{data}{data frame}

\item{maxit}{maximum number of iterations}
}
\value{
A list of
\item{coefficients}{Parameter estimates}
\item{fitted.values}{The model implied probabilities}
\item{residuals}{Dummy responses minus fitted values}
\item{pred.cat}{The model implied categories}
}
\description{
This function can be used to fit the full rank ordinal logistic regression model using
projected iterative weighted least squares.
}
\author{
David J. Hessen
}
