% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_safe.R
\name{solve_safe}
\alias{solve_safe}
\title{Solving Idx=u}
\usage{
solve_safe(I, u, lambda = 1e-08)
}
\arguments{
\item{I}{information matrix or Hessian}

\item{u}{score or gradient}

\item{lambda}{ridge penalty parameter}
}
\value{
ridge information matrix or Hessian
}
\description{
Solving Idx=u
}
