% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_se.R
\name{svd_se}
\alias{svd_se}
\title{Standard errors base on svd of Fisher information}
\usage{
svd_se(I, J, tol = NULL)
}
\arguments{
\item{I}{Fisher information matrix}

\item{J}{Jacobian matrix}

\item{tol}{tolerance}
}
\value{
standard errors
}
\description{
Standard errors base on svd of Fisher information
}
\author{
David J. Hessen
}
