% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_factor.R
\name{auto_factor}
\alias{auto_factor}
\title{Automatic recoding into factors}
\usage{
auto_factor(formula, data, max_levels = 10)
}
\arguments{
\item{formula}{model expression}

\item{data}{data frame}
}
\value{
recoded data frame
}
\description{
Automatic recoding into factors
}
