% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik_stn.R
\name{logLik_stn}
\alias{logLik_stn}
\title{The log-likelihood function}
\usage{
logLik_stn(x, X, Z, q)
}
\arguments{
\item{x}{parameter vector}

\item{X}{desgin matrix}

\item{Z}{response dummies}
}
\value{
log-likelihood value
}
\description{
The log-likelihood function
}
