% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Example dataset}
\format{
A data frame with 300 rows and 5 columns
\describe{
\item{y}{Response}
\item{X1 to X4}{Regressor variables}
}
}
\source{
Simulated data
}
\usage{
sim_data
}
\description{
A data set containing a single ordinal response variable and four continuous
regressor variables. The response variable has four levels.
}
\keyword{datasets}
