% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stn.R
\name{stn}
\alias{stn}
\title{Stereotype nominal regression}
\usage{
stn(formula, q = NULL, data, stv = NULL, csv = 0.001, tol = 1e-08, maxit = 100)
}
\arguments{
\item{formula}{model expression}

\item{q}{rank of the matrix of regressor effects. If not supplied, then it is set to its maximum.}

\item{data}{data frame}

\item{stv}{vector of starting values}

\item{csv}{common starting value}

\item{tol}{convergence tolerance}

\item{maxit}{maximum number of iterations}
}
\value{
A list of the following components
\item{alpha}{vector of estimated intercepts}
\item{Beta}{matrix of estimated regressor effects}
\item{Gamma, Phi}{matrices of parameter estimates}
\item{loglik}{value of the log-likelihood function}
\item{npars}{number of estimated parameters}
\item{AIC}{Akaike information criterion}
\item{BIC}{Bayesian information criterion}
\item{fitted.values}{model implied probabilities}
\item{residuals}{dummy responses minus fitted values}
\item{pred.cat}{model implied categories}
\item{obs.cat}{observed categories}
\item{CS, NK, MF}{Cox and Snell, Nagelkerke, and McFadden pseudo-R-squared values}
}
\description{
This function can be used to fit a q-dimensional stereotype nominal regression model.
}
\author{
David J. Hessen
}
