% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stn.R
\name{stn}
\alias{stn}
\title{Stereotype nominal regression}
\usage{
stn(formula, q = NULL, data, stv = NULL, csv = 1, tol = 1e-08, maxit = 100)
}
\arguments{
\item{formula}{model expression}

\item{q}{rank}

\item{data}{data frame}

\item{stv}{vector of starting values}

\item{csv}{common starting value}

\item{tol}{convergence tolerance}

\item{maxit}{maximum number of iterations}
}
\value{
A list of
\item{alpha, Beta, Gamma, Phi}{Parameter estimates}
\item{loglik}{Value of the log-likelihood function}
\item{npars}{Number of estimated parameters}
\item{AIC}{Akaike information criterion}
\item{BIC}{Bayesian information criterion}
\item{fitted.values}{Model implied probabilities}
\item{residuals}{Dummy responses minus fitted values}
\item{pred.cat}{Model implied categories}
\item{obs.cat}{Observed categories}
}
\description{
This function can be used to fit a q-dimensional stereotype nominal regression model.
}
\author{
David J. Hessen
}
