% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sto.R
\name{sto}
\alias{sto}
\title{Stereotype ordered regression}
\usage{
sto(formula, q = NULL, data, tol = 1e-08, maxit = 100)
}
\arguments{
\item{formula}{model expression}

\item{q}{rank}

\item{data}{data frame}

\item{tol}{convergence tolerance}

\item{maxit}{maximum number of iterations}
}
\value{
A list of
\item{alpha, Beta, Gamma, Phi}{Parameter estimates}
\item{loglik}{Value of the log-likelihood function}
\item{npars}{Number of estimated parameters}
\item{fitted.values}{Model implied probabilities}
\item{residuals}{Dummy responses minus fitted values}
\item{pred.cat}{Model implied categories}
\item{obs.cat}{Observed categories}
}
\description{
Stereotype ordered regression
}
