% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_se.R
\name{svd_se}
\alias{svd_se}
\title{Standard errors}
\usage{
svd_se(I, J, tol = NULL)
}
\arguments{
\item{I}{Fisher information matrix}

\item{J}{Jacobian matrix}

\item{tol}{tolerance}
}
\value{
standard errors
}
\description{
Robust standard errors obtained by svd of the Fisher information matrix
}
\author{
David J. Hessen
}
